# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [retail, wholesale, all]
        - in: query
          name: customerId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceDetail" }
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceDetail" }
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /invoices/{id}/payment:
    post:
      operationId: recordInvoicePayment
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceDetail" }
...
  /estimates:
    get:
      operationId: listEstimates
      tags: [estimates]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Estimate" }
    post:
      operationId: createEstimate
      tags: [estimates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EstimateInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EstimateDetail" }
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimates]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EstimateDetail" }
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /estimates/{id}/convert:
    post:
      operationId: convertEstimate
      tags: [estimates]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceDetail" }
...
components:
  schemas:
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    InvoiceItem:
      type: object
      properties:
        productId: { type: string }
        productName: { type: string }
        metal: { type: string }
        purity: { type: string }
        grossWeight: { type: number }
        lessWeight: { type: number }
        weightGrams: { type: number }
        wastagePercent: { type: number }
        ratePerGram: { type: number }
        makingChargePercent: { type: number }
        stoneCharges: { type: number }
        gstRate: { type: number }
        amount: { type: number }
      required: [productId, productName, metal, purity, grossWeight, lessWeight, weightGrams, wastagePercent, ratePerGram, makingChargePercent, stoneCharges, gstRate, amount]
    InvoiceItemInput:
      type: object
      properties:
        productId: { type: string }
        grossWeight: { type: number }
        lessWeight: { type: number }
        weightGrams: { type: number }
        wastagePercent: { type: number }
        ratePerGram: { type: number }
        makingChargePercent: { type: number }
        stoneCharges: { type: number }
        gstRate: { type: number }
      required: [productId, weightGrams, ratePerGram, makingChargePercent, stoneCharges, gstRate]
    OldGoldItem:
      type: object
      properties:
        description: { type: string }
        metal: { type: string, enum: [gold, silver] }
        grossWeight: { type: number }
        purityPercent: { type: number }
        fineWeight: { type: number }
        ratePerGram: { type: number }
        value: { type: number }
      required: [description, metal, grossWeight, purityPercent, fineWeight, ratePerGram, value]
    OldGoldItemInput:
      type: object
      properties:
        description: { type: string, minLength: 1 }
        metal: { type: string, enum: [gold, silver] }
        grossWeight: { type: number }
        purityPercent: { type: number }
        ratePerGram: { type: number }
      required: [description, metal, grossWeight, purityPercent, ratePerGram]
    Invoice:
      type: object
      properties:
        id: { type: string }
        invoiceNumber: { type: string }
        type: { type: string, enum: [retail, wholesale] }
        customerId: { type: string }
        customerName: { type: string }
        date: { type: string, format: date-time }
        subtotal: { type: number }
        gstAmount: { type: number }
        discount: { type: number }
        oldGoldValue: { type: number }
        total: { type: number }
        paidAmount: { type: number }
        balance: { type: number }
        status: { type: string, enum: [paid, partial, unpaid] }
        paymentMode: { type: string }
      required:
        [id, invoiceNumber, type, customerId, customerName, date, subtotal, gstAmount, discount, oldGoldValue, total, paidAmount, balance, status, paymentMode]
    InvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/Invoice"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/InvoiceItem" }
            oldGoldItems:
              type: array
              items: { $ref: "#/components/schemas/OldGoldItem" }
            notes: { type: string, nullable: true }
          required: [items, oldGoldItems]
    InvoiceInput:
      type: object
      properties:
        type: { type: string, enum: [retail, wholesale] }
        customerId: { type: string }
        date: { type: string, format: date-time }
        discount: { type: number }
        notes: { type: string, nullable: true }
        paymentMode: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItemInput" }
          minItems: 1
        oldGoldItems:
          type: array
          items: { $ref: "#/components/schemas/OldGoldItemInput" }
        paidAmount: { type: number }
      required: [type, customerId, date, discount, items, paidAmount]
    PaymentInput:
      type: object
      properties:
        amount: { type: number }
        note: { type: string, nullable: true }
      required: [amount]
...
    Estimate:
      type: object
      properties:
        id: { type: string }
        estimateNumber: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        date: { type: string, format: date-time }
        validUntil: { type: string, format: date-time, nullable: true }
        subtotal: { type: number }
        gstAmount: { type: number }
        discount: { type: number }
        total: { type: number }
        status: { type: string, enum: [draft, sent, accepted, converted, expired] }
        convertedInvoiceId: { type: string, nullable: true }
      required: [id, estimateNumber, customerId, customerName, date, subtotal, gstAmount, discount, total, status]
    EstimateDetail:
      allOf:
        - $ref: "#/components/schemas/Estimate"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/InvoiceItem" }
            notes: { type: string, nullable: true }
          required: [items]
    EstimateInput:
      type: object
      properties:
        customerId: { type: string }
        date: { type: string, format: date-time }
        validUntil: { type: string, format: date-time, nullable: true }
        discount: { type: number }
        notes: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItemInput" }
          minItems: 1
      required: [customerId, date, discount, items]
...