# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";
import type { InvoiceItem } from "./invoiceItem";
import type { OldGoldItem } from "./oldGoldItem";

export type InvoiceDetail = Invoice & {
  items: InvoiceItem[];
  oldGoldItems: OldGoldItem[];
  notes?: string | null;
};