# lib/api-zod/src/generated/types/oldGoldItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
 * OpenAPI spec version: 0.1.0
 */
import type { OldGoldItemInputMetal } from "./oldGoldItemInputMetal";

export interface OldGoldItemInput {
  /** @minLength 1 */
  description: string;
  metal: OldGoldItemInputMetal;
  grossWeight: number;
  purityPercent: number;
  ratePerGram: number;
}